añadido de las funciones de tratamientos

+++ programa/pacientes.cc
@@ -1,7 +1,7 @@
 #include "pacientes.h"
 
 Paciente::Paciente(string dni, string nombre, string apellidos, string fechanacimiento, string direccionpostal, 
-int telefono, string seguromutua){
+unsigned long int telefono, string seguromutua){
 
     dni_ = dni;
     nombre_ = nombre;
